move hf functions into hf_utils.py
