+++ distributed/meta_inference.sh
@@ -7,7 +7,7 @@ NGPU=${NGPU:-"2"}
 
 # TODO: We need to decide how to log for inference.
 # by default log just rank 0 output,
-LOG_RANK=${LOG_RANK:-0,1}
+LOG_RANK=${LOG_RANK:-0}
 
 overrides=""
 if [ $# -ne 0 ]; then
